test(allocation): split AllocationBucketQueryTest by query variants
Split the oversized bucket-query suite into month dimensions, time
granularity, hospital scope, and window/repository classes.
